+++ src/api-object.ts
@@ -1,10 +1,10 @@
 import { Construct, IConstruct } from 'constructs';
-import { resolve } from './_resolve';
 import { sanitizeValue } from './_util';
 import { App } from './app';
 import { Chart } from './chart';
 import { JsonPatch } from './json-patch';
 import { ApiObjectMetadata, ApiObjectMetadataDefinition } from './metadata';
+import { ResolutionContext } from './resolve';
 
 /**
  * Options for defining API objects.
@@ -197,16 +197,13 @@ export class ApiObject extends Construct {
    */
   public toJson(): any {
 
-    const resolver = App.of(this).resolver;
-    const apiObject = this;
-
     const data: any = {
       ...this.props,
       metadata: this.metadata.toJson(),
     };
 
     const sortKeys = process.env.CDK8S_DISABLE_SORT ? false : true;
-    const json = sanitizeValue(resolve([], data, apiObject, resolver), { sortKeys });
+    const json = sanitizeValue(this.resolve([], data), { sortKeys });
     const patched = JsonPatch.apply(json, ...this.patches);
 
     // reorder top-level keys so that we first have "apiVersion", "kind" and
@@ -221,6 +218,35 @@ export class ApiObject extends Construct {
 
     return result;
   }
+
+  private resolve(key: string[], value: any): any {
+
+    const resolvers = App.of(this).resolvers;
+
+    if (value == null) {
+      return value;
+    }
+
+    if (Array.isArray(value)) {
+      return value.map((x, i) => this.resolve([...key, `${i}`], x));
+    }
+
+    if (typeof(value) === 'object') {
+      const result: any = {};
+      for (const [k, v] of Object.entries(value)) {
+        result[k] = this.resolve([...key, k], v);
+      }
+      return result;
+    }
+
+    const context = new ResolutionContext(this, key, value);
+    for (const resolver of resolvers) {
+      resolver.resolve(context);
+      if (context.newValue !== value) return context.newValue;
+    }
+    return value;
+  }
+
 }
 
 function parseApiGroup(apiVersion: string) {
